% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{check_player_names}
\alias{check_player_names}
\title{Check for similar player names}
\usage{
check_player_names(x, distance_threshold = 4)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{read_dv}, or list of such objects}

\item{distance_threshold}{numeric: if two names differ by an amount less than this threshold, they will be returned as possible matches}
}
\value{
data.frame
}
\description{
Player names can sometimes be spelled incorrectly, particularly if there are character encoding issues. This can be a particular problem when combining data from multiple files. This function checks for similar names that might possibly be multiple variants on the same name.
}
\seealso{
\code{\link{read_dv}}, \code{\link{adist}}
}

