% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{create_meta_attacks}
\alias{create_meta_attacks}
\title{Create a meta attack data.frame from the plays object if it is missing}
\usage{
create_meta_attacks(plays)
}
\arguments{
\item{plays}{data.frame:}
}
\value{
a dataframe of attacks.
}
\description{
If your DataVolley files does not have a meta attack dataframe, 
(for example, if you are using Click&Scout), this function will take the
 plays object to create one..
}
\seealso{
\code{\link{read_dv}}
}
