% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{dv_attack_code2desc}
\alias{dv_attack_code2desc}
\title{Nominal descriptions for standard attack codes}
\usage{
dv_attack_code2desc(code)
}
\arguments{
\item{code}{character: vector of attack codes ("X5", "VP", etc)}
}
\value{
A named character vector of descriptions. Unrecognized attack codes will have \code{NA} description.
}
\description{
Nominal descriptions for standard attack codes
}
\examples{
dv_attack_code2desc(c("X5", "X7", "PP", "blah"))

}
