% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{dv_attack_code2loc}
\alias{dv_attack_code2loc}
\title{Nominal starting coordinate for standard attack codes}
\usage{
dv_attack_code2loc(code)
}
\arguments{
\item{code}{character: vector of attack codes ("X5", "VP", etc)}
}
\value{
A vector of numeric coordinates
}
\description{
Nominal starting coordinate for standard attack codes
}
\examples{
dv_attack_code2loc(code = c("X5", "X7", "PP"))

}
