% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{dv_attack_code_map}
\alias{dv_attack_code_map}
\title{Translate attack type and starting zone into an attack code.}
\usage{
dv_attack_code_map(type, start_zone)
}
\arguments{
\item{type}{string:}

\item{start_zone}{string:}
}
\value{
a vector of attack codes, set_types, etc.
}
\description{
If your DataVolley files does not have attack codes ready, (for example, if you are using Click&Scout), this function will take the starting zone and tempo of the attack to map it to an attack code.
}
\examples{
type = c("H", "Q", "T")
start_zone = c("8", "3", "4")
attack_code_map(type, start_zone)

}
\seealso{
\code{\link{read_dv}}
}
