% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{dv_attack_phase}
\alias{dv_attack_phase}
\title{Attack phase}
\usage{
dv_attack_phase(x)
}
\arguments{
\item{x}{datavolleyplays: the plays component of a datavolley object as returned by [dv_read()]}
}
\value{
Character vector
}
\description{
Attack phase as defined by DataVolley: either "Reception", "Transition sideout" or "Transition breakpoint", assigned only to attack actions.
}
