% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dv_combinations}
\alias{dv_combinations}
\title{Returns the attack combinations used during the scouting of the game}
\usage{
dv_combinations(x)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by read_dv}
}
\value{
A list of two tibbles. One with the scouting attack combinations, and one with the scouted combinations for each team during the game.
}
\description{
Returns the attack combinations used during the scouting of the game
}
\examples{
\dontrun{
library(ggplot2)
library(ggrepel)
library(dplyr)
x <- dv_read(dv_example_file())
dvc <- dv_combinations(x)

## Example 1: List of possible attack combinations 

ggplot(data= dvc$table_comb) + 
   ggcourt(court = "lower", labels = paste("Tempo: Tense"), court_colour = "indoor") + 
   geom_segment(aes(x = start_coordinate_x, y = start_coordinate_y, 
   xend = end_coordinate_x, yend = end_coordinate_y),
                         arrow = arrow(length = unit(0.2,"cm")), size=1) + 
   geom_text_repel(aes(x = start_coordinate_x, y = start_coordinate_y, label  = code), 
   size = 3, nudge_y = -0.25, segment.size = .1) + 
   facet_wrap(~ type)
   
## Example 2: List and frequency of attacks run by home team per setter position on reception phase

freq_comb <- count(dplyr::filter(plays(x), skill == "Attack", phase == "Reception", 
team == home_team(x)), 
attack_code, home_setter_position)

freq_join <- tidyr::drop_na(dplyr::left_join(dvc$table_comb, freq_comb, 
by = c("code" = "attack_code")), n)

ggplot(data = freq_join) + 
ggcourt(court = "lower", labels = "", court_colour = "indoor") + 
geom_segment(aes(x = start_coordinate_x, y = start_coordinate_y, 
xend = end_coordinate_x, yend = end_coordinate_y),
                         arrow = arrow(length = unit(0.2,"cm")), size=1)+
geom_text_repel(aes(x = start_coordinate_x, y = start_coordinate_y, label  = code), 
size = 3, nudge_y = -0.15, segment.size = .1) + 
geom_text_repel(aes(x = end_coordinate_x, y = 3.75, label  = n), 
size = 3, nudge_y = 0.5, direction = "x", segment.size = .1, vjust = 0) + 
facet_wrap(~ factor(home_setter_position, levels = c(4,3,2,5,6,1))) + 
ggtitle(paste(home_team(x), ": Offensive options on reception phase"))
}

}
