% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dv_cone_polygons}
\alias{dv_cone_polygons}
\title{The polygon coordinates for attack cones}
\usage{
dv_cone_polygons(zone, end = "upper", extended = FALSE)
}
\arguments{
\item{zone}{string: one of "L", "R", "M"}

\item{end}{string: use the "lower" or "upper" part of the figure}

\item{extended}{logical: if \code{FALSE}, the polygons will only cover the actual court area; if \code{TRUE}, they will be extended to cover the court periphery as well}
}
\value{
A data.frame with columns \code{cone_number}, \code{x}, \code{y}
}
\description{
The polygon coordinates for attack cones
}
\examples{
\dontrun{
 library(ggplot2)
 cxy <- dv_cone_polygons("M")
 ggplot(cxy, aes(x, y, group = cone_number, fill = as.factor(cone_number))) +
   geom_polygon() + ggcourt()
}
}
