% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{dv_court}
\alias{dv_court}
\title{Plot a volleyball court diagram}
\usage{
dv_court(
  plot_package = "base",
  court = "full",
  show_zones = TRUE,
  labels = c("Serving team", "Receiving team"),
  as_for_serve = FALSE,
  show_zone_lines = TRUE,
  show_minor_zones = FALSE,
  grid_colour = "black",
  zone_colour = "grey70",
  minor_zone_colour = "grey80",
  fixed_aspect_ratio = TRUE,
  zone_font_size = 10,
  ...
)
}
\arguments{
\item{plot_package}{string: either "base" or "ggplot2". If "ggplot2", the \code{\link{ggcourt}} function is used}

\item{court}{string: "full" (show full court) or "lower" or "upper" (show only the lower or upper half of the court)}

\item{show_zones}{logical: add numbers indicating the court zones (3m squares)?}

\item{labels}{string: labels for the lower and upper court halves (pass NULL for no labels)}

\item{as_for_serve}{logical: if TRUE and \code{show_zones} is TRUE, show zones as for serving. Only zones 1,5,6,7,9 are meaningful in this case}

\item{show_zone_lines}{logical: if FALSE, just show the 3m line. If TRUE, also show the 3m x 3m zones}

\item{show_minor_zones}{logical: add lines for the subzones (1.5m squares)?}

\item{grid_colour}{string: colour to use for court sidelines, 3m line, and net}

\item{zone_colour}{string: colour to use for zone lines and labels}

\item{minor_zone_colour}{string: colour to use for minor zone grid lines}

\item{fixed_aspect_ratio}{logical: if TRUE, coerce the plotted court to be square (for a half-court plot) or a 2:1 rectangle (full court plot). Prior to package version 0.5.3 this was not TRUE by default}

\item{zone_font_size}{numeric: the font size of the zone labels}

\item{...}{: additional parameters passed to \code{ggplot2::theme_classic(...)}}
}
\description{
Volleyball court schematic suitable for adding to a figure
}
\details{
The datavolley package uses the following dimensions and coordinates for plotting:
\itemize{
  \item the court is shown such that the sidelines are oriented vertically and the net is oriented horizontally
  \item the intersection of the left-hand sideline and the bottom baseline is at (0.5, 0.5)
  \item the intersection of the right-hand sideline and the top baseline is at (3.5, 6.5)
  \item the net intersects the sidelines at (0.5, 3.5) and (3.5, 3.5)
  \item the zones 1-9 (as defined in the DataVolley manual) on the lower half of the court are located at:
    \enumerate{
      \item (3, 1)
      \item (3, 3)
      \item (2, 3)
      \item (1, 3)
      \item (1, 1)
      \item (2, 1)
      \item (1, 2)
      \item (2, 2)
      \item (3, 2)
    }
  \item the zones 1-9 (as defined in the DataVolley manual) on the upper half of the court are located at:
    \enumerate{
      \item (1, 6)
      \item (1, 4)
      \item (2, 4)
      \item (3, 4)
      \item (3, 6)
      \item (2, 6)
      \item (3, 5)
      \item (2, 5)
      \item (1, 5)
    }
}

To get a visual depiction of this, try: \code{ ggplot() + ggcourt() + theme_bw()}
}
\examples{
\dontrun{
x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)

library(dplyr)

## Example: attack frequency by zone, per team

attack_rate <- plays(x) \%>\% dplyr::filter(skill == "Attack") \%>\%
  group_by(team, start_zone) \%>\% dplyr::summarize(n_attacks = n()) \%>\%
  mutate(rate = n_attacks/sum(n_attacks)) \%>\% ungroup

## add columns "x" and "y" for the x,y coordinates associated with the zones
attack_rate <- cbind(attack_rate, dv_xy(attack_rate$start_zone, end = "lower"))

## for team 2, these need to be on the top half of the diagram
tm2 <- attack_rate$team == teams(x)[2]
attack_rate[tm2, c("x", "y")] <- dv_xy(attack_rate$start_zone, end = "upper")[tm2, ]

## plot it
dv_heatmap(attack_rate[, c("x", "y", "rate")], legend_title = "Attack rate")

## add the court diagram
dv_court(labels = teams(x))
}

}
\seealso{
\code{\link{ggcourt}} for a \code{ggplot2} equivalent function; \code{\link{dv_xy}}, \code{\link{dv_xy2index}}, \code{\link{dv_index2xy}}, \code{\link{dv_flip_xy}}
}
