% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{dv_example_file}
\alias{dv_example_file}
\title{Example DataVolley files provided as part of the datavolley package}
\usage{
dv_example_file(choice = 1)
}
\arguments{
\item{choice}{numeric: which data file to return?
\itemize{
  \item{1 - the January 2015 Slovenian junior women's final between Braslovče and Nova KBM Branik}
  \item{2 - the December 2012 men's Slovenian national championship semifinal between ACH Volley and Maribor}
  \item{3 - Nicaragua vs Cuba women from the Pan American Cup, August 2022 (vsm format, courtesy Christophe Elek)}
}}
}
\value{
path to the file
}
\description{
Example DataVolley files provided as part of the datavolley package
}
\examples{
myfile <- dv_example_file()
x <- dv_read(myfile, insert_technical_timeouts = FALSE)
summary(x)

}
\references{
The example data files came from \url{http://www.odbojka.si/}
}
\seealso{
\code{\link{dv_read}}
}
