% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dv_find_to_flip_coordinates}
\alias{dv_find_to_flip_coordinates}
\title{Find coordinates that need flipping}
\usage{
dv_find_to_flip_coordinates(x, target_start_end = "lower")
}
\arguments{
\item{x}{datavolleyplays: the plays component of a datavolley object as returned by \code{dv_read}}

\item{target_start_end}{string: "lower" or "upper"}
}
\value{
A logical index with length equal to the number of rows of \code{x}. TRUE indicates rows of \code{x} that need their coordinates flipped
}
\description{
The orientation of coordinates (e.g. is a serve going from the lower part of the court to the upper, or vice-versa?) depends on how the scout entered them. This function finds coordinates that require flipping, so that all attacks/serves/whatever can be plotted with the same orientation
}
\seealso{
\code{\link{dv_flip_xy}}
}
