% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dv_index2xy}
\alias{dv_index2xy}
\alias{dv_xy2index}
\title{Grid index to x,y coordinate and vice-versa}
\usage{
dv_index2xy(index)

dv_xy2index(x, y)
}
\arguments{
\item{index}{integer: vector of grid indices. If missing, the entire grid will be returned. The row numbers match the grid indices}

\item{x}{numeric: x-coordinate. For \code{dv_index2xy} this can be a two-column matrix or data.frame containing x and y}

\item{y}{numeric: y-coordinate}
}
\value{
for dv_index2xy, a data.frame with columns "x" and "y"; for dv_xy2index a vector of integer values
}
\description{
DataVolley uses a grid to represent positions on court (values in columns "start_coordinate", "mid_coordinate", and "end_coordinate" in the play-by-play data frame). These functions convert grid index values to x, y coordinates suitable for plotting, and vice-versa. For a description of the court dimensons and coordinates see \code{\link{ggcourt}}.
}
\examples{

## positions (zones) 1 and 3 are at x, y coordinates c(3, 1) and c(2, 3) respectively

## their grid indices:
dv_xy2index(c(3, 2), c(1, 3))

}
\seealso{
\code{\link{ggcourt}}, \code{\link{dv_xy}}, \code{\link{dv_flip_xy}}, \code{\link{dv_xy2zone}}, \code{\link{dv_xy2subzone}}
}
