% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dv_int2rgb}
\alias{dv_int2rgb}
\alias{dv_rgb2int}
\title{Convert integer colour to RGB}
\usage{
dv_int2rgb(z)

dv_rgb2int(x)
}
\arguments{
\item{z}{integer: vector of integers}

\item{x}{integer: vector of hex colour strings}
}
\value{
Character vector of hex RGB colour strings
}
\description{
DataVolley files use an integer representation of colours. These functions convert to and from hex colour strings as used in R.
}
\examples{
dv_int2rgb(c(255, 16711680))

}
