% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_set_codes.R
\name{dv_meta_setter_calls}
\alias{dv_meta_setter_calls}
\title{Extract the setter calls table from a datavolley object}
\usage{
dv_meta_setter_calls(x)
}
\arguments{
\item{x}{datavolley: as returned by \code{\link[=dv_read]{dv_read()}}}
}
\value{
The setter calls table from the metadata section of \code{x}
}
\description{
Extract the setter calls table from a datavolley object
}
\details{
Note that some columns are placeholders (named e.g. \code{X2}, \code{X4} and generally filled with NAs). These are kept for compatibility reasons.
}
\examples{
x <- dv_read(dv_example_file())

## the setter calls in this file
dv_meta_setter_calls(x)

}
