% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{dv_meta_video}
\alias{dv_meta_video}
\alias{dv_meta_video<-}
\title{Get or set the video metadata in a datavolley object}
\usage{
dv_meta_video(x)

dv_meta_video(x) <- value
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by [datavolley::dv_read()]}

\item{value}{string or data.frame: a string containing the path to the video file, or a data.frame with columns "camera" and "file"}
}
\value{
For `dv_meta_video`, the existing video metadata. For `dv_meta_video<-`, the video metadata value in `x` is changed
}
\description{
Get or set the video metadata in a datavolley object
}
\examples{
x <- dv_read(dv_example_file())
dv_meta_video(x) ## empty dataframe
dv_meta_video(x) <- "/path/to/my/videofile"
dv_meta_video(x)

}
