% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{dv_point_phase}
\alias{dv_point_phase}
\title{Point phase}
\usage{
dv_point_phase(x)
}
\arguments{
\item{x}{datavolleyplays: the plays component of a datavolley object as returned by [dv_read()]}
}
\value{
Character vector
}
\description{
Point phase as defined by DataVolley: either "Sideout" or "Breakpoint", assigned only to winning or losing actions (including green codes). Note that the point phase is inferred for the winning action (i.e. the point phase value for both the winning and losing action is "Sideout" if the winning team was receiving).
}
