% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.R
\name{dv_read_sq}
\alias{dv_read_sq}
\title{Read a team roster (*.sq) file}
\usage{
dv_read_sq(
  filename,
  do_transliterate = FALSE,
  encoding = "guess",
  date_format = "guess",
  surname_case = "asis",
  verbose = FALSE
)
}
\arguments{
\item{filename}{string: file name to read}

\item{do_transliterate}{logical: should we transliterate all text to ASCII?}

\item{encoding}{character: text encoding to use. Text is converted from this encoding to UTF-8. A vector of multiple encodings can be provided, and this function will attempt to choose the best. If encoding is "guess", the encoding will be guessed}

\item{date_format}{string: the expected date format (used for dates of birth). One of "ymd", "mdy", "dmy", or "guess". If \code{date_format} is something other than "guess", that date format will be preferred where dates are ambiguous}

\item{surname_case}{string or function: should we change the case of player surnames? If \code{surname_case} is a string, valid values are "upper","lower","title", or "asis"; otherwise \code{surname_case} may be a function that will be applied to the player surname strings}

\item{verbose}{logical: if \code{TRUE}, show progress}
}
\value{
A list with two components: "team" and "players", both of which are data frames
}
\description{
Read a team roster (*.sq) file
}
\examples{
\dontrun{
  x <- dv_read_sq("/path/to/my/roster_file")
}
}
