% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_set_codes.R
\name{dv_remap_attack_codes}
\alias{dv_remap_attack_codes}
\title{Remap attack codes to new values}
\usage{
dv_remap_attack_codes(x, from_codes, to_codes, allow_remove = FALSE)
}
\arguments{
\item{x}{datavolley: as returned by \code{\link[=dv_read]{dv_read()}}}

\item{from_codes}{character: one or more attack codes to change}

\item{to_codes}{character: one or more replacement attack codes. \code{to_codes} must either be the same length as \code{from_codes} (in which case each element of \code{from_codes} will be changed to its corresponding element in \code{to_codes}), or \code{to_codes} must be a single string (in which case each element of \code{from_codes} will be changed to this code)}

\item{allow_remove}{logical: if a given value of \code{to_code} is NA or an empty string (""), the corresponding \code{from_code} will be removed, so long as \code{allow_remove} is TRUE. It is not entirely clear how useful this removal functionality is, so the extra \code{allow_remove} parameter must be set in order to avoid inadvertant removals}
}
\value{
A modified copy of \code{x}
}
\description{
Remap attack codes to new values
}
\details{
If a given value of \code{to_code} does not exist in the attacks metadata table in \code{x}, we will just rename \code{from_code} to \code{to_code} without changing any other details. If \code{to_code} does exist in the attacks metadata table, we will rename the code and also change the attack description, attack and set tempo (\code{skill_type} column) and set target (\code{set_type}), if those details are different for the \code{to_code}. The \code{start_zone} is not updated in either case.

If a given value of \code{to_code} is NA or an empty string (""), the corresponding \code{from_code} will be removed, so long as \code{allow_remove} is TRUE. It is not entirely clear how useful this removal functionality is, so the extra \code{allow_remove} parameter must be set in order to avoid inadvertant removals.
}
\examples{
x <- dv_read(dv_example_file())

## the attack codes in this file
dv_meta_attack_codes(x)

## and their usage in the play-by-play data
table(plays(x)$attack_code)

## Example 1
## relabel V5 and V6 as Z5 and Z6
## (leaving their other details unchanged, because Z5 and Z6
##  do not exist in the attack codes table)
x2 <- dv_remap_attack_codes(x, from_codes = c("V5", "V6"), to_codes = c("Z5", "Z6"))

## check that the table has changed
dv_meta_attack_codes(x2)
## and their usage in the play-by-play data
table(plays(x2)$attack_code)

## Example 2
## relabel "C5" and "JJ" to "X5" (they are all fast sets to position 4)
ac <- dv_meta_attack_codes(x)
ac[ac$code \%in\% c("X5", "C5", "JJ"), ]

x2 <- dv_remap_attack_codes(x, from_codes = c("C5", "JJ"), to_codes = "X5")
ac2 <- dv_meta_attack_codes(x2)
ac2[ac2$code \%in\% c("X5", "C5", "JJ"), ]
table(plays(x2)$attack_code)

}
