% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.R
\name{dv_repair}
\alias{dv_repair}
\title{Attempt to repair a datavolley object}
\usage{
dv_repair(x)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by [dv_read()]}
}
\value{
A modified copy of `x`. If problems exist and cannot be repaired, an error will be thrown
}
\description{
Currently an attempt will be made to repair these issues:
* if multiple players on the same team have the same jersey number, players with that number (on that team) who did not take to the court will be removed from their team roster. In this situation, whether or not a player took to the court is determined from the match metadata only
* if multiple players have the same player ID but different jersey numbers, players with that ID who did not take to the court will be removed from their team roster. In this situation, whether or not a player took to the court is determined from the match metadata and the play-by-play data
}
