% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{dv_sync_summary}
\alias{dv_sync_summary}
\title{Summarize the video sync times in a dvw file}
\usage{
dv_sync_summary(x)
}
\arguments{
\item{x}{datavolley: a single datavolley object as returned by \code{\link{dv_read}}, or the \code{plays} component of one}
}
\value{
A data.frame with columns \code{type}, \code{N}, \code{mean}, \code{most_common}, \code{min}, \code{max}
}
\description{
This function will generate a summary of various video time differences in a dvw file. Apply this to a file that you have synchronized to video, and the results can be used to tweak the behaviour of \code{\link{dv_sync_video}}.
}
\examples{
x <- dv_read(dv_example_file(3))
dv_sync_summary(x)

}
\seealso{
\code{\link{dv_sync_video}}
}
