% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dv_xy2subzone}
\alias{dv_xy2subzone}
\title{Convert x, y coordinates to zones and subzones}
\usage{
dv_xy2subzone(x, y = NULL)
}
\arguments{
\item{x}{numeric: the x coordinate}

\item{y}{numeric: the y coordinate. If \code{y} is \code{NULL}, \code{x} will be treated as a grid index (see \code{\link{dv_index2xy}})}
}
\value{
A tibble with columns \code{zone} and \code{subzone}
}
\description{
Convert x, y coordinates to zones and subzones
}
\examples{
\dontrun{

## a bunch of random points on and around the court
idx <- round(runif(100, min = 1, max = 10000))

## convert to zones
zn <- dv_xy2subzone(x = idx)

## or, equivalently, convert the index to xy values first
zn <- cbind(zn, dv_index2xy(idx))

## plot
ggplot(zn, aes(x, y, colour = as.factor(zone), shape = subzone)) + geom_point(size = 3) +
  ggcourt(labels = NULL)

## the points shoud be coloured by zone
}

}
\seealso{
\code{\link{dv_xy2index}}, \code{\link{dv_index2xy}}, \code{\link{dv_cone2xy}}, \code{\link{dv_xy2zone}}
}
