% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dv_xy2zone}
\alias{dv_xy2zone}
\title{Convert x, y coordinates to zones}
\usage{
dv_xy2zone(x, y = NULL, as_for_serve = FALSE)
}
\arguments{
\item{x}{numeric: the x coordinate}

\item{y}{numeric: the y coordinate. If \code{y} is \code{NULL}, \code{x} will be treated as a grid index (see \code{\link{dv_index2xy}})}

\item{as_for_serve}{logical: if \code{TRUE}, treat the zones as if they refer to serving locations (i.e. zone 7 in between zones 5 and 6, and zone 9 in between zones 6 and 1)}
}
\value{
A numeric vector giving the zone number
}
\description{
Convert x, y coordinates to zones
}
\examples{
\dontrun{

## a bunch of random points on and around the court
idx <- round(runif(100, min = 1, max = 10000))

## convert to zones
zn <- dv_xy2zone(x = idx)

## or, equivalently, convert the index to xy values first
idx_xy <- dv_index2xy(idx)
zn <- dv_xy2zone(x = idx_xy$x, idx_xy$y)

## plot
ggplot(idx_xy, aes(x, y, fill = as.factor(zn))) + geom_point(shape = 21) +
  ggcourt(labels = NULL)

## the points shoud be coloured by zone
}

}
\seealso{
\code{\link{dv_xy2index}}, \code{\link{dv_index2xy}}, \code{\link{dv_cone2xy}}, \code{\link{dv_xy2subzone}}
}
