% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{dvlist_summary}
\alias{dvlist_summary}
\alias{print.summary.datavolleylist}
\title{Summarize a list of volleyball matches}
\usage{
dvlist_summary(z)

\method{print}{summary.datavolleylist}(x, ...)
}
\arguments{
\item{z}{list: list of datavolley objects as returned by read_dv()}
}
\value{
named list with various summary indicators, including a competition ladder
}
\description{
Summarize a list of volleyball matches
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  summary(list(x,x)) ## same match duplicated twice, just for illustration purposes
}

}
\seealso{
\code{\link{read_dv}}
}

