% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{find_first_attack}
\alias{find_first_attack}
\title{Find first attacks by the receiving team (i.e. attacks associated with a serve reception)}
\usage{
find_first_attack(x)
}
\arguments{
\item{x}{data.frame: the plays component of a datavolley object, as returned by \code{read_dv()}}
}
\value{
named list with components "ix" (logical indices into the x object where the row corresponds to a first attack in a rally), "n" (number of receptions for which there was a first attack by the receiving team), "n_win" (the number of winning first attacks), "win_rate" (number of winning first attacks as a proportion of the total number of first attacks).
}
\description{
Find first attacks by the receiving team (i.e. attacks associated with a serve reception)
}
\examples{
\dontrun{
x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
## first attack win rate, by team
by(x$plays,x$plays$team,function(z)find_first_attack(z)$win_rate)
}
}
\seealso{
\code{\link{read_dv}}
}

