% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_match}
\alias{find_match}
\title{Find a particular match in a list of datavolley objects}
\usage{
find_match(match_id, x)
}
\arguments{
\item{match_id}{string: match_id to find}

\item{x}{list: list of datavolley objects as returned by \code{dv_read}}
}
\value{
numeric index of the match in the list
}
\description{
Find a particular match in a list of datavolley objects
}
\seealso{
\code{\link{dv_read}}
}
