% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{find_player_name_remapping}
\alias{find_player_name_remapping}
\title{Attempt to build a player name remapping table}
\usage{
find_player_name_remapping(x, distance_threshold = 3, verbose = TRUE)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{read_dv}, or list of such objects}

\item{distance_threshold}{numeric: if two names differ by an amount less than this threshold, they will be treated as the same name}

\item{verbose}{logical: print progress to console as we go? Note that warnings will also be issued regardless of this setting}
}
\value{
data.frame with columns team, from, to
}
\description{
A player name can sometimes be spelled incorrectly, particularly if there are character encoding issues. This can be a particular problem when combining data from multiple files. This function will attempt to find names that have been misspelled and create a remapping table suitable to pass to \code{\link{remap_player_names}}. Player names will only be compared within the same team. Note that this function is unlikely to get perfect results: use its output with care.
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  ## artifically change one player name to be a mis-spelled version of
  ##  another player's name, just for demo purposes
  x <- remap_player_names(x,data.frame(team="Br.Maribor",from="Iza MLAKAR",
                            to="Marina KASCIC"))
  remap <- find_player_name_remapping(x)
}
}
\seealso{
\code{\link{remap_player_names}}, \code{\link{check_player_names}}
}

