% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{find_runs}
\alias{find_runs}
\title{Generate information about runs of events}
\usage{
find_runs(x, idvars = "team", within_set = TRUE)
}
\arguments{
\item{x}{data.frame: a subset of the plays component of a datavolley object, as returned by \code{read_dv()}}

\item{idvars}{character: string or character vector of variabe names to use to identify the entity doing the events}

\item{within_set}{logical: only consider runs within a single set? If FALSE, runs that span sets will be treated as a single run}
}
\value{
A data.frame the same number of rows as \code{x}, and with columns \code{run_id} (the identifier of the run to which each row belongs), \code{run_length} (the length of the run), and \code{run_position} (the position of this row in its associated run).
}
\description{
Find runs of events within a match. Typically, this function would be passed a subset of \code{plays(x)}, such as rows
corresponding to serves. Runs that are terminated by the end of a set are not assigned a \code{run_length}.
}
\examples{
\dontrun{
## find runs of serves
x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
serve_idx <- find_serves(plays(x))
serve_run_info <- find_runs(plays(x)[serve_idx,])
## distribution of serve run lengths
table(unique(serve_run_info[,c("run_id","run_length")])$run_length)
}

}
\seealso{
\code{\link{read_dv}} \code{\link{plays}}
}

