% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{find_serves}
\alias{find_serves}
\title{Find serves}
\usage{
find_serves(x)
}
\arguments{
\item{x}{data.frame: the plays component of a datavolley object, as returned by \code{read_dv()}}
}
\value{
a logical vector, giving the indices of the rows of x that correspond to serves
}
\description{
Find serves
}
\examples{
\dontrun{
x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
  insert_technical_timeouts=FALSE)
serve_idx <- find_serves(plays(x))
## number of serves by team
table(plays(x)$team[serve_idx])
}

}
\seealso{
\code{\link{read_dv}} \code{\link{plays}}
}

