% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{findnext}
\alias{findnext}
\title{Find each entry in y that follows each entry in x}
\usage{
findnext(x, y)
}
\arguments{
\item{x}{numeric: vector}

\item{y}{numeric: vector}
}
\value{
vector, each entry is the value in y that is next-largest to each corresponding entry in x
}
\description{
Find each entry in y that follows each entry in x
}
\examples{
findnext(c(1,5,10),c(1,2,3,7,8,9))

}
