% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{fix_ace_evaluations}
\alias{fix_ace_evaluations}
\title{Find aces that might not be marked as such}
\usage{
fix_ace_evaluations(x, verbose = TRUE)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{read_dv}, or list of such objects}

\item{verbose}{logical: print progress to console?}
}
\value{
datavolley object or list of such with updated evaluation values
}
\description{
Some DataVolley files do not indicate serve aces with the skill evaluation "Ace". This function will search for winning serves, either with no reception or a reception error, and change their evaluation value to "Ace"
}
\seealso{
\code{\link{read_dv}}
}

