% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggcourt}
\alias{ggcourt}
\title{Add volleyball court schematic to a ggplot}
\usage{
ggcourt(court = "full", show_zones = TRUE, show_labels = TRUE)
}
\arguments{
\item{court}{string: "full" (show full court) or "attack" or "defence" (show only the attacking or defending half of the court)}

\item{show_zones}{logical: add numbers indicating the court zones?}

\item{show_labels}{logical: add labels indicating the attacking and receiving sides of the court?}
}
\value{
ggplot layer
}
\description{
Add volleyball court schematic to a ggplot
}
\examples{
\dontrun{
x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
    insert_technical_timeouts=FALSE)

## calculate attack frequency by zone
attack_rate <- as.data.frame(xtabs(~start_zone,data=subset(plays(x),skill=="Attack")),
    stringsAsFactors=FALSE)
attack_rate$start_zone <- as.numeric(attack_rate$start_zone)
attack_rate$rate <- attack_rate$Freq/sum(attack_rate$Freq)

## plot
attack_rate <- cbind(attack_rate,ggxy(attack_rate$start_zone,type="start"))
ggplot(attack_rate,aes(x,y,fill=rate))+geom_tile()+ggcourt()+
    scale_fill_gradient2(name="Attack rate")
}

\dontrun{
## show map of starting and ending zones of attacks
## tabulate attacks by starting and ending zone
attack_rate <- as.data.frame(xtabs(~start_zone+end_zone,data=subset(plays(x),skill=="Attack")),
    stringsAsFactors=FALSE)
attack_rate$start_zone <- as.numeric(attack_rate$start_zone)
attack_rate$end_zone <- as.numeric(attack_rate$end_zone)
attack_rate$rate <- attack_rate$Freq/sum(attack_rate$Freq)
attack_rate <- attack_rate[attack_rate$Freq>0,]
## starting x,y coordinates
temp <- ggxy(attack_rate$start_zone,type="start")
names(temp) <- c("sx","sy")
attack_rate <- cbind(attack_rate,temp)
## ending x,y coordinates
temp <- ggxy(attack_rate$end_zone,type="end")
names(temp) <- c("ex","ey")
attack_rate <- cbind(attack_rate,temp)
## plot in reverse order so largest arrows are on the bottom
attack_rate <- attack_rate[order(attack_rate$rate,decreasing=TRUE),]
p <- ggplot(attack_rate,aes(x,y,col=rate))+ggcourt()+scale_fill_gradient2(name="Attack rate")
for (n in 1:nrow(attack_rate))
    p <- p+geom_path(data=data.frame(
            x=c(attack_rate$sx[n],attack_rate$ex[n]),
            y=c(attack_rate$sy[n],attack_rate$ey[n]),
            rate=attack_rate$rate[n]),
        aes(size=rate),lineend="round",arrow=arrow(ends="last",type="closed"))
p+scale_fill_gradient(name="Attack rate")+guides(size="none")
}
}
\seealso{
\code{\link{ggxy}}
}

