% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggxy-deprecated}
\alias{ggxy-deprecated}
\title{Court zones to x, y coordinates}
\usage{
ggxy(zones, end="lower", xynames=c("x", "y"), as_for_serve=FALSE)
}
\arguments{
\item{zones}{numeric: zones numbers 1-9 to convert to x and y coordinates}

\item{end}{string: use the "lower" or "upper" part of the figure}

\item{xynames}{character: names to use for the x and y columns of the returned data.frame}

\item{as_for_serve}{logical: if TRUE, treat positions as for serving. Only zones 1,5,6,7,9 are meaningful in this case}
}
\value{
data.frame with columns "x" and "y" (or other names if specified in \code{xynames})
}
\description{
Generate x and y coordinates for plotting, from DataVolley numbered zones
}
\seealso{
\code{\link{datavolley-deprecated}}
}
\keyword{internal}
