% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggxy}
\alias{ggxy}
\title{Create x and y coordinates for plotting, from DataVolley start/end zones}
\usage{
ggxy(zones, type = "start")
}
\arguments{
\item{zones}{numeric: zones numbers 1-9 to convert to x and y coordinates}

\item{type}{string: "start" or "end" zones. Start zones are plotted on the lower part of the figure, end zones on the upper}
}
\value{
data.frame with x and y components
}
\description{
Create x and y coordinates for plotting, from DataVolley start/end zones
}
\examples{
\dontrun{
x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
    insert_technical_timeouts=FALSE)

## calculate attack frequency by zone
attack_rate <- as.data.frame(xtabs(~start_zone,data=subset(plays(x),skill=="Attack")),
    stringsAsFactors=FALSE)
attack_rate$start_zone <- as.numeric(attack_rate$start_zone)
attack_rate$rate <- attack_rate$Freq/sum(attack_rate$Freq)

## plot
attack_rate <- cbind(attack_rate,ggxy(attack_rate$start_zone,type="start"))
ggplot(attack_rate,aes(x,y,fill=rate))+geom_tile()+ggcourt()+
    scale_fill_gradient2(name="Attack rate")
}
}
\seealso{
\code{\link{ggcourt}}
}

