% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{meta}
\alias{meta}
\alias{meta<-}
\title{Extract the meta component from a datavolley object, or assign a new one}
\usage{
meta(x)

meta(x) <- value
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{dv_read}}

\item{value}{list: new data}
}
\value{
The meta component of x (a list), or a modified version of x with the new meta component inserted
}
\description{
Extract the meta component from a datavolley object, or assign a new one
}
\examples{
\dontrun{
  x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
  str(meta(x))

  m <- meta(x)
  m$match$date <- Sys.Date() ## change the match date
  meta(x) <- m
}
}
\seealso{
\code{\link{dv_read}}
}
