% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{plays}
\alias{plays}
\title{Extract the plays component from a datavolley object}
\usage{
plays(x)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{dv_read}}
}
\value{
The plays component of x (a data.frame)
}
\description{
Extract the plays component from a datavolley object
}
\examples{
\dontrun{
  x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
  inspect(plays(x))
}
}
\seealso{
\code{\link{dv_read}}
}
