% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{print.summary.datavolley}
\alias{print.summary.datavolley}
\title{Print method for summary.datavolley}
\usage{
\method{print}{summary.datavolley}(x, ...)
}
\arguments{
\item{x}{summary.datavolley: a summary.datavolley object as returned by \code{summary.datavolley}}

\item{...}{: additional arguments (currently these have no effect)}
}
\description{
Print method for summary.datavolley
}
\seealso{
\code{\link{summary.datavolley}}
}
