% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{print.summary.datavolleylist}
\alias{print.summary.datavolleylist}
\title{Print method for summary.datavolleylist}
\usage{
\method{print}{summary.datavolleylist}(x, ...)
}
\arguments{
\item{x}{summary.datavolleylist: a summary.datavolleylist object, as returned by \code{dvlist_summary}}

\item{...}{: additional arguments (currently these have no effect)}
}
\description{
Print method for summary.datavolleylist
}
\seealso{
\code{\link{dvlist_summary}}
}
