% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dv_insert_sets}
\alias{dv_insert_sets_check}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{vscoututils}{\code{\link[vscoututils]{dv_insert_sets}}, \code{\link[vscoututils]{dv_insert_sets_check}}}
}}

