% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_names.R
\name{remap_player_info}
\alias{remap_player_info}
\title{Change player information}
\usage{
remap_player_info(x, remap)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{dv_read}, or list of such objects}

\item{remap}{data.frame: data.frame of strings with columns team, name_from, and any of player_id, firstname, and lastname}
}
\value{
A datavolley object or list with corresponding player names changed
}
\description{
An experimental function to replace \code{remap_player_names} as a more comprehensive remapping of player attributes.
}
\examples{
\dontrun{
  x <- dv_read(dv_example_file(), insert_technical_timeouts = FALSE)
  x <- remap_player_info(x, data.frame(team = c("Nova KBM Branik", "BraslovÄe"),
                                       name_from = c("ELA PINTAR", "KATJA MIHALINEC"),
                                       firstname = c("Ela", "Katja"), stringsAsFactors = FALSE))
}

}
