% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remap_player_names}
\alias{remap_player_names}
\title{Change player names}
\usage{
remap_player_names(x, remap)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{read_dv}, or list of such objects}

\item{remap}{data.frame: data.frame of strings with columns team, from, and to}
}
\value{
datavolley object or list with corresponding player names changed
}
\description{
A player name can sometimes be spelled incorrectly, particularly if there are character encoding issues. This can be a particular problem when combining data from multiple files. A player matching the \code{team} and \code{from} entries in a row in \code{remap} is renamed to the corresponding \code{to} value.
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  x <- remap_player_names(x,data.frame(team=c("Kamnik","Br.Maribor"),
                            from=c("Ana Katarina HRIBAR","Marina CVETANOVIC"),
                            to=c("Ana-Katarina HRIBAR","Marina CVETANOVIĆ")))
}
}
\seealso{
\code{\link{read_dv}}, \code{\link{check_player_names}}
}

