% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{remap_team_names}
\alias{remap_team_names}
\title{Change team names}
\usage{
remap_team_names(x, from, to)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{read_dv}, or list of such objects}

\item{from}{string: vector of team names to change from}

\item{to}{string: vector of team names to change to}
}
\value{
datavolley object with corresponding team names changed
}
\description{
A team name can sometimes be spelled incorrectly, particularly if there are character encoding issues. This can be a particular problem when combining data from multiple files. Any team name in \code{to} is replaced with the corresponding entry in \code{from}.
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  summary(x)
  x <- remap_team_names(x,from="Kamnik",to="OK Kamnik")
  summary(x)
}
}
\seealso{
\code{\link{read_dv}}
}

