% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotations.R
\name{rotations}
\alias{rotations}
\title{Get team rotations point by point}
\usage{
rotations(x, team, point_ids, new_rotation = NULL)
}
\arguments{
\item{x}{datavolley or data.frame: a datavolley object as returned by \code{read_dv}, or the plays component of that object}

\item{team}{string: team name}

\item{point_ids}{vector: vector of pointids for which to return the rotation}

\item{new_rotation}{vector: vector of player number, positionned from 1 to 6.}
}
\value{
list of 2 data.frames. Current data.frame, without changes, and updated data.frame, with new player rotation.
}
\description{
Get team rotations point by point
}
\examples{
\dontrun{
  x <- read_dv(dv_example_file(), insert_technical_timeouts=FALSE)
  rotations(x)
}
}
\seealso{
\code{\link{read_dv}}
}
