% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{serve_win_points}
\alias{serve_win_points}
\title{Find serve win points}
\usage{
serve_win_points(x, return_id = FALSE)
}
\arguments{
\item{x}{data.frame: the plays component of a datavolley object, as returned by \code{read_dv()}}

\item{return_id}{logical: include the match_id and point_id of all serve win points in the returned object?}
}
\value{
named list with components "ix" (logical indices of serves corresponding to serve win points in the x object), "n" (number of serve win points in x), "rate" (serve win rate from x). If \code{return_id} is TRUE, also return a component "id" (a data.frame containing the match_id and point_id of all serve win points)
}
\description{
Find points in which the serving team wins the point. Serve win rate is the fraction of serves won by the serving team.
}
\examples{
\dontrun{
x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
serve_idx <- find_serves(x$plays)
swp <- serve_win_points(x$plays)
## number of serves by team
table(x$plays$team[serve_idx])
## number of points won on serve by team
table(x$plays$team[serve_idx & swp$ix])
}

}
\seealso{
\code{\link{read_dv}}
}

