% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{skill_evaluation_decoder}
\alias{skill_evaluation_decoder}
\title{Translate skill evaluation codes into meaningful summary phrases}
\usage{
skill_evaluation_decoder(skill, evaluation_code, show_map = FALSE)
}
\arguments{
\item{skill}{string: one-character skill code, either "S"erve, "R"eception, "A"ttack, "B"lock, "D"ig, s"E"t, or "F"reeball}

\item{evaluation_code}{string: one-character evaluation code, generally one of =/-+#!}

\item{show_map}{logical: if TRUE, return the whole table being used to map evaluation codes to summary phrases}
}
\value{
string giving the interpretation of that skill evaluation code
}
\description{
If your DataVolley files use evaluation codes differently to those coded here, you will need to supply a custom
skill_evaluation_decode function to \code{\link{read_dv}}
}
\examples{
skill_evaluation_decoder("S","#")
skill_evaluation_decoder(show_map=TRUE)

}
\seealso{
\code{\link{read_dv}}
}

