% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{summary.datavolley}
\alias{summary.datavolley}
\title{Print a simple summary of a datavolley object}
\usage{
\method{summary}{datavolley}(object, ...)
}
\arguments{
\item{object}{datavolley: datavolley object as returned by read_dv()}

\item{...}{: additional arguments (currently these have no effect)}
}
\value{
print summary to console, and return the summary object invisibly
}
\description{
Print a simple summary of a datavolley object
}
\seealso{
\code{\link{read_dv}}
}

