% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{teams}
\alias{teams}
\title{Get team names from datavolley object}
\usage{
teams(x)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{read_dv}}
}
\value{
character vector of team names
}
\description{
Get team names from datavolley object
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
    insert_technical_timeouts=FALSE)
  teams(x)
}
}
\seealso{
\code{\link{read_dv}}
}

