% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_dv}
\alias{validate_dv}
\title{Additional validation checks on a datavolley object}
\usage{
validate_dv(x)
}
\arguments{
\item{x}{datavolley: datavolley object as returned by \code{read_dv}}
}
\value{
data.frame
}
\description{
This function is automatically run as part of \code{read_dv} if \code{do_extra_validation} is TRUE
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
    insert_technical_timeouts=FALSE)
  xv <- validate_dv(x)
}

}
\seealso{
\code{\link{read_dv}}
}

