% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{break_points}
\alias{break_points}
\title{Calculate break points}
\usage{
break_points(x)
}
\arguments{
\item{x}{data.frame: the plays component of a datavolley object, as returned by \code{read_dv()}}
}
\value{
named list with components "ix" (logical indices of break points in the x object), "n" (number of break points in x), "rate" (break point rate from x)
}
\description{
A break point in one in which the serving team wins the point. Break point rate is the fraction of points won by the serving team.
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  break_points(subset(x$plays,team=="Kamnik"))
}

}
\seealso{
\code{\link{read_dv}}
}

