% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavolley.R
\docType{package}
\name{datavolley}
\alias{datavolley}
\alias{datavolley-package}
\title{\pkg{datavolley}}
\description{
The example data provided with the datavolley package came from \url{http://www.odbojka.si/}, and is the January 2015 junior women's final between Braslovče and Nova KBM Branik.
}
\author{
Ben Raymond \email{ben@untan.gl}
}

