% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavolley.R
\docType{package}
\name{datavolley}
\alias{datavolley}
\alias{datavolley-package}
\title{\pkg{datavolley}}
\description{
Provides basic functions for parsing Datavolley scouting files. Datavolley is a software package used for scouting and summarizing volleyball matches.
}
\details{
The example data files provided with the datavolley package came from \url{http://www.odbojka.si/}:
\itemize{
  \item{\code{system.file("extdata/example_data.dvw",package="datavolley")}: the January 2015 junior women's final between Braslovče and Nova KBM Branik}
  \item{\code{system.file("extdata/PM06.dvw",package="datavolley")}: the December 2012 men's Slovenian national championship semifinal between ACH Volley and Maribor}
}
}
\author{
Ben Raymond \email{ben@untan.gl}
}
