% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavolley.R
\docType{package}
\name{datavolley}
\alias{datavolley}
\title{\pkg{datavolley}}
\description{
Provides basic functions for parsing Datavolley scout files. Datavolley is a software package used for scouting and summarizing volleyball matches.
}
\details{
The example data files provided with the datavolley package came from \url{http://www.odbojka.si/}.
}
\author{
Ben Raymond \email{ben@untan.gl}
}
