% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{dv_attack_code_map}
\alias{dv_attack_code_map}
\title{Translate attack type and starting zone into an attack code.}
\usage{
dv_attack_code_map(type, start_zone)
}
\arguments{
\item{type}{character: vector of attack tempos ("H", "T", "Q", etc). A \code{type} vector of length 1 will be expanded to the length of the \code{start_zone} vector, if needed}

\item{start_zone}{integer: vector of start zones}
}
\value{
A vector of attack codes, set_types, etc.
}
\description{
If your DataVolley files does not have attack codes ready, (for example, if you are using Click&Scout), this function will take the starting zone and tempo of the attack to map it to an attack code.
}
\examples{
dv_attack_code_map(type = c("H", "Q", "T"), start_zone = c("8", "3", "4"))

}
