% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{dv_sync_video}
\alias{dv_sync_video}
\alias{dv_sync_contact_times}
\alias{dv_sync_offsets}
\title{Synchronize video times}
\usage{
dv_sync_video(
  x,
  first_serve_contact,
  freeball_dig_time_offset = NA,
  contact_times = dv_sync_contact_times(),
  offsets = dv_sync_offsets(),
  times_from,
  enforce_order = TRUE
)

dv_sync_contact_times(...)

dv_sync_offsets(...)
}
\arguments{
\item{x}{datavolley: a single datavolley object as returned by \code{\link{dv_read}}}

\item{first_serve_contact}{numeric or string: the video time of the first serve contact. This can be a numeric value giving the time in seconds from the start of the video, or a string of the form "MM:SS" (minutes and seconds) or "HH:MM:SS" (hours, minutes and seconds)}

\item{freeball_dig_time_offset}{numeric: if non-NA, the clock times of freeball digs will be used directly in the synchronization process. Freeball digs will be aligned using their clock times relative to the first serve contact clock time, with this \code{freeball_dig_time_offset} value (in seconds) added. So if when scouting live you typically enter freeball digs one second after they happen, use \code{freeball_dig_time_offset = -1}. If \code{freeball_dig_time_offset} is NA, which is the default, the clock times of freeball digs will not be used in the synchronization process}

\item{contact_times}{list: a set of parameters that control the synchronization process. See Details, below}

\item{offsets}{list: a list set of offsets to be added to each contact time in the second step of the synchronization process. See Details, below. If \code{offsets} is NULL or an empty list, no offsets are applied}

\item{times_from}{string: either "clock" or "video": take the serve times (and freeball dig times, if `freeball_dig_time_offset` is non-NA) from clock or video times. By default, clock times are used unless they are all missing}

\item{enforce_order}{logical: the estimated contact times will always be time-ordered (the contact time of a given touch cannot be prior to the contact time of a preceding touch). But the offsets can be different for different skills, leading to final video times that are not time ordered. These will be fixed if \code{enforce_order} is TRUE}

\item{...}{: name-value pairs of elements to override the defaults in \code{dv_sync_contact_times} and \code{dv_sync_offsets}}
}
\value{
A copy of \code{x} with modified \code{video_time} values in its \code{plays} component
}
\description{
This function uses the time of each serve and some rules to align the other contacts in a rally with their (approximately correct) times in the corresponding match video. Warning: experimental!
}
\details{
When a match is scouted live, the clock time of each serve will usually be correct because the scout can enter the serve code at the actual time of serve. But the remainder of the touches in the rally might not be at their correct times if the scout can't keep up with the live action. This function makes some assumptions about typical contact-to-contact times to better synchronize the scouted contacts with the corresponding match video.

The clock time of each serve will be used as the reference time for each rally. If clock times are not present in the file, the video time of each serve will be used instead. If those are also missing, the function will fail.

Freeball digs can optionally be treated in the same way as serves, with their scouted times used directly in the synchronization process. Obviously this only makes sense if the scout has actually been consistent in their timing when entering freeball digs, but assuming that is the case then setting the \code{freeball_dig_time_offset} to a non-NA value will improve the synchronization of rallies with freeballs. These rallies otherwise tend to synchronize poorly, because the play is messy and less predictable compared to in-system rallies.

Note that this synchronization relies on the serve clock times in the file being consistent, and so it will only work if the match has been scouted in a single sitting (either live, or from video playback but without pausing/rewinding/fast-forwarding the video). If your clock times are not consistent but the video time of each serve is correct, then you can use the video time of each serve as the reference time instead.

The synchronization is a two-step process. In the first step, the video time of each scouted contact is estimated (i.e. the actual time that the player made contact with the ball). In the second step, skill-specific offsets are added to those contact times. (This is important if your video montage software uses the synchronized video times directly, because you will normally want a video clip to start some seconds before the actual contact of interest).

The \code{contact_times} object contains a set of times (in seconds), which you can adjust to suit your scouting style and level of play. If you have an already-synchronized dvw file, the \code{\link{dv_sync_summary}} function can provide some guidance as to what these values should be. The \code{contact_times} object contains the following entries:
\itemize{
  \item SQ - time between the scouted serve time and actual serve contact for jump serves
  \item SM - time between the scouted serve time and actual serve contact for jump-float serves
  \item SO - time between the scouted serve time and actual serve contact for all other serves
  \item SQ_R, SM_R, SO_R - the time between serve contact and reception contact for jump, jump-float, and other serves
  \item R_E - the time between reception contact and set contact
  \item EQ_A - the time between set contact and attack contact for quick sets
  \item EH_A - the time between set contact and attack contact for high sets
  \item EO_A - the time between set contact and attack contact for all other sets
  \item A_B - the time between attack contact and block contact
  \item A_D - the time between attack contact and dig contact (no intervening block touch)
  \item A_B_D - the time between attack contact and dig contact (with block touch)
  \item D_E - the time between dig contact and set contact
  \item RDov - the time between reception or dig overpass contact and the next touch by the opposition
  \item END - the time between the last contact and end-of-rally marker
}

The \code{offsets} object defines the offset (in seconds) to be added to each contact time in the second pass of the synchronization process. It contains the entries "S" (serve), "R" (reception), "E" (set), "A" (attack), "D", (dig), "B" (block), and "F" (freeball).

Note that the entries in \code{contact_times} and \code{offsets} can be fractions. The actual video time entries in the returned file are required to be integers and so the final values will be rounded, but using fractional values (particularly for the \code{contact_times} entries) can give better accuracy in the intermediate calculations.
}
\examples{
x <- dv_read(dv_example_file())
## first serve contact was at 54s in the video
x <- dv_sync_video(x, first_serve_contact = 54)

## with a custom configuration
my_contact_times <- dv_sync_contact_times(SQ = 3) ## override default entries as necessary
## first serve contact was at 3:35 in the video
x <- dv_sync_video(x, first_serve_contact = "3:35", contact_times = my_contact_times)

}
\seealso{
\code{\link{dv_sync_summary}}
}
