% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{dvlist_summary}
\alias{dvlist_summary}
\title{Summarize a list of volleyball matches}
\usage{
dvlist_summary(z)
}
\arguments{
\item{z}{list: list of datavolley objects as returned by \code{dv_read}}
}
\value{
named list with various summary indicators, including a competition ladder
}
\description{
Summarize a list of volleyball matches
}
\examples{
\dontrun{
  x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
  dvlist_summary(list(x,x)) ## same match duplicated twice, just for illustration purposes
}

}
\seealso{
\code{\link{dv_read}}
}
