% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{find_first_attack}
\alias{find_first_attack}
\title{Find first attacks by the receiving team (i.e. attacks associated with a serve reception)}
\usage{
find_first_attack(x)
}
\arguments{
\item{x}{data.frame: the plays component of a datavolley object, as returned by \code{dv_read()}}
}
\value{
named list with components "ix" (logical indices into the x object where the row corresponds to a first attack in a rally), "n" (number of receptions for which there was a first attack by the receiving team), "n_win" (the number of winning first attacks), "win_rate" (number of winning first attacks as a proportion of the total number of first attacks).
}
\description{
Find first attacks by the receiving team (i.e. attacks associated with a serve reception)
}
\examples{
\dontrun{
x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
## first attack win rate, by team
by(plays(x),plays(x)$team,function(z)find_first_attack(z)$win_rate)
}
}
\seealso{
\code{\link{dv_read}} \code{\link{plays}}
}
