% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_plays.R
\name{fix_ace_evaluations}
\alias{fix_ace_evaluations}
\title{Find aces that might not be marked as such}
\usage{
fix_ace_evaluations(x, rotation_error_is_ace = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{dv_read}, or list of such objects}

\item{rotation_error_is_ace}{logical: should a rotation error on reception by the receiving team be counted as an ace?}

\item{verbose}{logical: print progress to console?}
}
\value{
datavolley object or list of such with updated evaluation values
}
\description{
Some DataVolley files do not indicate serve aces with the skill evaluation "Ace". This function will search for winning serves, either with no reception or a reception error, and change their evaluation value to "Ace"
}
\seealso{
\code{\link{dv_read}}
}
