% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{inspect}
\alias{inspect}
\title{Convenience function for inspecting the plays component of a datavolley object}
\usage{
inspect(x, vars = "minimal", maxrows = 100, extra)
}
\arguments{
\item{x}{datavolleyplays: the plays component of a datavolley object as returned by \code{read_dv}}

\item{vars}{string: which variables to print? "minimal" set or "all"}

\item{maxrows}{numeric: maximum number of rows to print}

\item{extra}{character: names of any extra columns to include in the output}
}
\description{
Convenience function for inspecting the plays component of a datavolley object
}
\examples{
\dontrun{
  x <- read_dv(dv_example_file(), insert_technical_timeouts=FALSE)
  inspect(plays(x))
}

}
\seealso{
\code{\link{read_dv}} \code{\link{plays}}
}
