% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{play_phase}
\alias{play_phase}
\title{Figure out the phase of play associated with each point}
\usage{
play_phase(x)
}
\arguments{
\item{x}{datavolleyplays: the plays component of a datavolley object as returned by \code{read_dv}}
}
\value{
character vector
}
\description{
Phase is either "Serve", "Reception" (serve reception and the attack immediately following it),
or "Transition" (play not associated with serve or reception)
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
    insert_technical_timeouts=FALSE)
  px <- plays(x)
  px$phase <- play_phase(px)
}

}
\seealso{
\code{\link{read_dv}} \code{\link{plays}}
}
