% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{read_dv}
\alias{read_dv}
\title{Read a datavolley file}
\usage{
read_dv(filename, dowarn = TRUE, encodings = c("latin2", "windows-1252",
  "iso885913"))
}
\arguments{
\item{filename}{string: file name to read}

\item{dowarn}{logical: should we issue warnings about the contents of the file as we read it?}

\item{encodings}{character: vector of text encodings to try}
}
\value{
named list with meta and plays components. meta provides match metadata, plays is the main point-by-point data in the form of a data.frame
}
\description{
Read a datavolley file
}
\examples{
\dontrun{

}
}
\references{
\url{http://www.dataproject.com/IT/en/Volleyball}
}

