% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{read_dv}
\alias{read_dv}
\title{Read a datavolley file}
\usage{
read_dv(filename, insert_technical_timeouts = TRUE, dowarn = FALSE,
  encodings = c("latin2", "windows-1252", "iso885913"))
}
\arguments{
\item{filename}{string: file name to read}

\item{insert_technical_timeouts}{logical: should we insert technical timeouts at points 8 and 16 of sets 1--4?}

\item{dowarn}{logical: should we issue warnings about the contents of the file as we read it?}

\item{encodings}{character: vector of text encodings to try}
}
\value{
named list with meta and plays components. meta provides match metadata, plays is the main point-by-point data in the form of a data.frame
}
\description{
Read a datavolley file
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  summary(x)
}
}
\references{
\url{http://www.dataproject.com/IT/en/Volleyball}
}

