% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{read_dv}
\alias{read_dv}
\title{Read a datavolley file}
\usage{
read_dv(filename, insert_technical_timeouts = TRUE, do_warn = FALSE,
  do_transliterate = FALSE, encodings = c("latin2", "windows-1252",
  "iso885913"))
}
\arguments{
\item{filename}{string: file name to read}

\item{insert_technical_timeouts}{logical: should we insert technical timeouts at points 8 and 16 of sets 1--4?}

\item{do_warn}{logical: should we issue warnings about the contents of the file as we read it?}

\item{do_transliterate}{logical: should we transliterate all text to ASCII? See details}

\item{encodings}{character: vector of text encodings to try. Text encoding is converted to UTF-8 internally}
}
\value{
named list with \code{meta} and \code{plays} components. \code{meta} provides match metadata, \code{plays} is the main point-by-point data in the form of a data.frame
}
\description{
The \code{do_transliterate} option may be helpful when trying to work with multiple files from the same competition, since different text encodings may be used on different files. This can lead to e.g. multiple versions of the same team name. Transliterating can help avoid this, at the cost of losing e.g. diacriticals. Transliteration is applied after converting from the guessed/specified text encoding to UTF-8.
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"))
  summary(x)
}
}
\references{
\url{http://www.dataproject.com/IT/en/Volleyball}
}

