% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_names.R
\name{remap_team_names}
\alias{remap_team_names}
\title{Change team names}
\usage{
remap_team_names(x, remap, fixed = TRUE)
}
\arguments{
\item{x}{datavolley: a datavolley object as returned by \code{dv_read}, or list of such objects}

\item{remap}{data.frame: data.frame of strings with one or both columns \code{from} and \code{team_id}, and one or both columns \code{to} and \code{to_team_id}}

\item{fixed}{logical: treat the \code{from} and \code{team_id} entries as fixed strings? If \code{fixed} is \code{FALSE} they will be treated as regular expressions}
}
\value{
datavolley object or list with corresponding team names changed
}
\description{
A team name can sometimes be spelled incorrectly, particularly if there are character encoding issues. This can be a particular problem when combining data from multiple files. If a team name matches the \code{from} entry and/or its ID matches the \code{team_id} entry in a row in \code{remap}, the team will be renamed to the corresponding \code{to} value and/or its ID changed to the corresponding \code{to_team_id} value.
}
\examples{
\dontrun{
  x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
  summary(x)

  ## rename a team based just on team name
  summary(remap_team_names(x, data.frame(from="Nova KBM Branik", to="NKBM Branik")))

  ## rename a team based on team name and ID
  summary(remap_team_names(x, data.frame(from="Nova KBM Branik", to="NKBM Branik", team_id="MB4")))
}
}
\seealso{
\code{\link{dv_read}}
}
