% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{serve_win_points}
\alias{serve_win_points}
\title{Find serve win points}
\usage{
serve_win_points(x, return_id = FALSE)
}
\arguments{
\item{x}{data.frame: the plays component of a datavolley object, as returned by \code{dv_read()}}

\item{return_id}{logical: include the match_id and point_id of all serve win points in the returned object?}
}
\value{
named list with components "ix" (logical indices of serves corresponding to serve win points in the x object), "n" (number of serve win points in x), "rate" (serve win rate from x). If \code{return_id} is TRUE, also return a component "id" (a data.frame containing the match_id and point_id of all serve win points)
}
\description{
Find points in which the serving team wins the point. Serve win rate is the fraction of serves won by the serving team.
}
\examples{
\dontrun{
x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
serve_idx <- find_serves(plays(x))
swp <- serve_win_points(plays(x))
## number of serves by team
table(plays(x)$team[serve_idx])
## number of points won on serve by team
table(plays(x)$team[serve_idx & swp$ix])
}

}
\seealso{
\code{\link{dv_read}} \code{\link{plays}}
}
