% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{skill_evaluation_decoder}
\alias{skill_evaluation_decoder}
\title{Translate skill evaluation codes into meaningful summary phrases}
\usage{
skill_evaluation_decoder(style = "default")
}
\arguments{
\item{style}{string: currently "default" (following the standard definitions described in the DataVolley manual) or "volleymetrics" (per the conventions that VolleyMetrics use)}
}
\value{
function. This function takes arguments skill, evaluation_code, and show_map and returns a string giving the interpretation of that skill evaluation code
}
\description{
If your DataVolley files use evaluation codes differently to those coded here, you will need to supply a custom
skill_evaluation_decode function to \code{\link{dv_read}}
}
\examples{
sd <- skill_evaluation_decoder()
sd("S","#")
sd(show_map=TRUE)

}
\seealso{
\code{\link{dv_read}}
}
