% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{summary.datavolley}
\alias{summary.datavolley}
\title{A simple summary of a volleyball match}
\usage{
\method{summary}{datavolley}(object, ...)
}
\arguments{
\item{object}{datavolley: datavolley object as returned by \code{read_dv}}

\item{...}{: additional arguments (currently these have no effect)}
}
\value{
list of summary items
}
\description{
A simple summary of a volleyball match
}
\examples{
\dontrun{
  x <- read_dv(system.file("extdata/example_data.dvw",package="datavolley"),
    insert_technical_timeouts=FALSE)
  summary(x)
}

}
\seealso{
\code{\link{read_dv}}
}
