% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dv.R
\name{summary.datavolley}
\alias{summary.datavolley}
\title{A simple summary of a volleyball match}
\usage{
\method{summary}{datavolley}(object, ...)
}
\arguments{
\item{object}{datavolley: datavolley object as returned by \code{dv_read}}

\item{...}{: additional arguments (currently these have no effect)}
}
\value{
list of summary items
}
\description{
A simple summary of a volleyball match
}
\examples{
x <- dv_read(dv_example_file(), insert_technical_timeouts=FALSE)
summary(x)

}
\seealso{
\code{\link{dv_read}}
}
